Set winner when player leaves elimination mid-game
Some players immediately leave the game mid-way in elimination maps
when they have the final turn in the team and they are about to
die/fall off a cliff, which causes the game to draw when the opponent
team should have been declared the winner.

The game will now check if the player to leave the elimination game was
the last player in the team and in case they were so, mark their team
score so the opponent team is declared as the winner.