Fixed the CTF Flag Bug
There was an issue on how the flag return was handled.
When the player who at the time was returning the flag died, the counter wouldn't stop and the flag would still return to the base. 
Funny part? This bug would persist through the entire match. The only way to temporarily fix it was to throw the flag out of bounds.

I provided a simple fix, so now when the player who returns the flag dies, the flag won't return by itself.